refactor(clawdstrike-policy-event): extract supply_chain_cli from edr/mod.rs
Move the package-manager + cloud CLI classification cluster
(command_looks_like_package_manager, package_registry_cli_name,
suspicious_package_registry_cli_reason, package_registry_auth_config_
reference, package_registry_credential_env_keys, cloud_cli_name,
suspicious_cloud_cli_reason, cloud_credential_env_keys) into
edr/supply_chain_cli.rs.

detection/supply_chain.rs reaches seven of these via super::super::, so
they are re-exported pub(crate) at the edr root (the cluster's private
fns were widened to pub(crate) only for those seven; package_registry_
auth_config_reference stays private). Pure relocation; bodies unchanged.
